add ios arm64 support, also enabling several compiler warning to flag up implicit conversion issues. Buddy: Malcolm

+++ ios/ios_src/GlDisplayService.mm
@@ -25,22 +25,22 @@ bool GlDisplayService::IsDisplayAvailable() const
 	return m_displayBound;
 }
 
-int GlDisplayService::GetDisplayWidth() const
+float GlDisplayService::GetDisplayWidth() const
 {
 	return m_displayWidth;
 }
 
-int GlDisplayService::GetDisplayHeight() const
+float GlDisplayService::GetDisplayHeight() const
 {
 	return m_displayHeight;
 }
 
-int GlDisplayService::GetDisplayDpi() const
+float GlDisplayService::GetDisplayDpi() const
 {
 	return m_displayDpi;
 }
 
-int GlDisplayService::GetPixelScale() const
+float GlDisplayService::GetPixelScale() const
 {
 	return m_pixelScale;
 }
@@ -88,7 +88,7 @@ bool GlDisplayService::TryBindDisplay(GLKView& view)
 	UIScreen* screen = [UIScreen mainScreen];
 	if ([screen respondsToSelector: @selector(scale)])
 	{
-		m_pixelScale = screen.scale;
+		m_pixelScale = static_cast<float>(screen.scale);
 	}
 	else
 	{
@@ -124,8 +124,8 @@ bool GlDisplayService::TryBindDisplay(GLKView& view)
 		std::swap(screenBounds.width, screenBounds.height);
 	}
 
-	m_displayWidth = screenBounds.width * m_pixelScale;
-	m_displayHeight = screenBounds.height * m_pixelScale;
+	m_displayWidth = static_cast<float>(screenBounds.width) * m_pixelScale;
+	m_displayHeight = static_cast<float>(screenBounds.height) * m_pixelScale;
 	m_displayDpi = App::GetDeviceDpi() * m_pixelScale;
 
 	m_displayBound = true;
